2009-07-10 14:42 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * source/rtl/gtwvt/gtwvt.c     * Added one parenthesis (sync with GTWVG).
  * contrib/gtwvg/gtwvg.c
  * contrib/gtwvg/gtwvg.h
  * contrib/gtwvg/wvggui.c
  * contrib/gtwvg/wvgutils.c
  * contrib/gtwvg/wvgcore.c
    * Got rid of [U]SHORTs.

  * examples/terminal/trm_cli.prg
  * examples/terminal/terminal.prg
  * examples/terminal/trm_srv.prg
  * examples/terminal/terminal.hbc
  * examples/terminal/trm_misc.c
    % Using hbwin lib function instead of local implementations.

+++ harbour/contrib/gtwvg/wvggui.c
@@ -2114,7 +2114,7 @@ static BOOL hb_gt_wvt_Info( PHB_GT pGT, int iType, PHB_GT_INFO pInfo )
 
                      pWVT->pfnLayered( pWVT->hWnd,
                                        RGB( 255,255,255 ),
-                                       hb_itemGetNI( pInfo->pNewVal2 ),
+                                       ( BYTE ) hb_itemGetNI( pInfo->pNewVal2 ),
                                        /*LWA_COLORKEY|*/ LWA_ALPHA );
                   }
 #endif
